Update component folder structure. Rename .spec. to .test.
