+++ src/components/ActionMenu/index.ts
@@ -1,0 +1,2 @@
+import ActionMenu from './ActionMenu';
+export default ActionMenu;
