+++ src/components/AddNewMenu/index.ts
@@ -1,0 +1,2 @@
+import AddNewMenu from './AddNewMenu';
+export default AddNewMenu;
