+++ src/components/AuthenticationWrapper/index.ts
@@ -1,0 +1,2 @@
+import AuthenticationWrapper from './AuthenticationWrapper';
+export default AuthenticationWrapper;
