+++ src/components/CircleProgress/index.ts
@@ -1,0 +1,2 @@
+import CircleProgress from './CircleProgress';
+export default CircleProgress;
