+++ src/components/ErrorState/index.ts
@@ -1,0 +1,2 @@
+import ErrorState from './ErrorState';
+export default ErrorState;
