+++ src/components/PrimaryNav/index.ts
@@ -1,0 +1,2 @@
+import PrimaryNav from './PrimaryNav';
+export default PrimaryNav;
