+++ src/components/PromiseLoader/index.ts
@@ -1,0 +1,2 @@
+import PromiseLoader from './PromiseLoader';
+export default PromiseLoader;
