+++ src/components/ShowMore/index.ts
@@ -1,0 +1,2 @@
+import ShowMore from './ShowMore';
+export default ShowMore;
