+++ src/components/UserMenu/index.ts
@@ -1,0 +1,2 @@
+import UserMenu from './UserMenu';
+export default UserMenu;
