+++ src/components/WithDocumentation/index.ts
@@ -1,0 +1,2 @@
+import WithDocumentation from './WithDocumentation';
+export default WithDocumentation;
