+++ src/features/linodes/LinodesLanding/LinodeRow.tsx
@@ -12,7 +12,7 @@ import Typography from 'material-ui/Typography';
 import TableRow from 'material-ui/Table/TableRow';
 import TableCell from 'material-ui/Table/TableCell';
 
-import ActionMenu, { Action } from 'src/components/ActionMenu';
+import ActionMenu, { Action } from 'src/components/ActionMenu/ActionMenu';
 import LinodeStatusIndicator from 'src/components/LinodeStatusIndicator';
 import Tag from 'src/components/Tag';
 import RegionIndicator from './RegionIndicator';
